Adding a docker compose setup; Adding a dragonfly package; Adding derived add_auth_spec
