+++ examples/key-value-store.yaml
@@ -1,0 +1,21 @@
+name: my_store
+module: Service
+endpoints:
+  - name: peaches
+    module: KeyValueStore
+    calibration: '2*{}'
+    initial_value: 0.75
+    log_interval: 10
+    get_on_set: True
+    log_on_set: True
+  - name: chips
+    module: KeyValueStore
+    calibration: 'times3({})'
+    initial_value: 1.75
+  - name: waffles
+    module: KeyValueStore
+    #log_interval: 30
+    #log_on_set: True
+    calibration: '1.*{}'
+    initial_value: 4.00
+
