+++ setup.py
@@ -1,6 +1,6 @@
 from setuptools import setup, find_namespace_packages
 
-packages = find_namespace_packages('.', include=['dripline.extensions.*'])
+packages = find_namespace_packages('.', include=['dragonfly', 'dripline.extensions', 'dripline.extensions.*'])
 print('packages are: {}'.format(packages))
 
 setup(
